Created test dir and split tests.py
