+++ ticketing/tests/signing.py
@@ -1,83 +1,12 @@
-import json
-import rsa
 import base64
+from unittest import TestCase
 
-from django.conf import settings
+import rsa
 from django.contrib.auth.models import User
-from django.test import TestCase
-from rest_framework.test import APIRequestFactory, force_authenticate
-from ticketing.api.userviews import NotificationView
 
-from ticketing.models import Account, BalanceTicket, Driver
+from tessera import settings
+from ticketing.models import Account, BalanceTicket
 from ticketing.userticket.createqrcode import QRCode
-from ticketing.driverauth.driverauthtoken import DriverAuth, BadTokenError
-
-
-class DriverIdTestCase(TestCase):
-
-    def setUp(self):
-        self.driver = Driver(name="Bobby Tables", pin=1234)
-        self.driver.save()
-
-    def test_create_token(self):
-        auth = DriverAuth(secret_key="thisisatest")
-        token = auth.createtoken(self.driver.id)
-        if token != "MS45LU03aF9sUVVDVEZYVDRDOGZ5WFF0TlhYMVk=":
-            self.fail("Token doesn't match expected. Got " + token +
-                      " expected MS45LU03aF9sUVVDVEZYVDRDOGZ5WFF0TlhYMVk=")
-
-    def test_verify_token_pass(self):
-        auth = DriverAuth(secret_key="thisisatest")
-        try:
-            verify = auth.verifytoken("MS45LU03aF9sUVVDVEZYV"
-                                      "DRDOGZ5WFF0TlhYMVk=")
-            if verify.id != self.driver.id:
-                self.fail("Driver id is different expected: " + verify.id +
-                          " got: " + self.driver.id)
-            else:
-                pass
-        except BadTokenError:
-            self.fail("Object does not exist")
-
-    def test_very_token_fail(self):
-        auth = DriverAuth(secret_key="thisisiatest")
-        try:
-            verify = auth.verifytoken("MS43cjJGQXB4SjZOcmZsb01fQj"
-                                      "UteWY0MC1PeXX=")
-
-            if verify.id is self.driver.id:
-                pass
-            # should only reach here if verify function is incorrect
-            self.fail("Exec continued")
-        except BadTokenError:
-            pass
-
-
-class CreateNotificationTestCase(TestCase):
-
-    def setUp(self):
-        self.user = User.objects.create_user("fakeuser", "fakepassword")
-        Account(user=self.user).save()
-        self.acc = self.user.account.all()[0]
-        self.factory = APIRequestFactory()
-        self.view = NotificationView.as_view()
-
-    def test_create_notification(self):
-        req = self.factory.post('/api/v1/users/notifications/',
-                                {"token": "1234"}, format='json')
-
-        force_authenticate(req, user=self.user)
-
-        resp = self.view(req).rendered_content
-        resp_dict = json.loads(resp.decode('UTF-8'))
-
-        self.assertEqual(resp_dict["success"], True)
-        self.assertIsNotNone(resp_dict["created_id"])
-
-        push = self.acc.notification_tokens.all()[0]
-
-        self.assertIsNotNone(push)
-        self.assertEqual(resp_dict['created_id'], push.id)
 
 
 class SignTestCase(TestCase):
